+++ week4/Parktaehee/src/Home.css
@@ -176,7 +176,7 @@ button {
   width: 48px;
   height: 48px;
   background-color: #ff7e36;
-  left: 323px;
+  left: 331px;
   top: 643px;
   border-radius: 100%;
   color: white;
